python/requirements.txt
pyyaml==3.12
Jinja2==2.8
